Integrate new jobs feed
