+++ commercial/app/controllers/commercial/JobAds.scala
@@ -1,8 +1,7 @@
 package controllers.commercial
 
 import play.api.mvc._
-import scala.util.Random
-import model.commercial.jobs.{LightJobsAgent, JobsAgent}
+import model.commercial.jobs.JobsAgent
 import common.{JsonNotFound, JsonComponent}
 import model.Cached
 
@@ -10,25 +9,11 @@ object JobAds extends Controller {
 
   def jobs = Action {
     implicit request =>
-      JobsAgent.matchingAds(segment) match {
+      JobsAgent.adsTargetedAt(segment) match {
         case Nil => JsonNotFound.apply()
         case jobs => {
-          val shuffled = Random.shuffle(jobs)
-          Cached(60)(JsonComponent(views.html.jobs(shuffled take 5)))
+          Cached(60)(JsonComponent(views.html.jobs(jobs take 5)))
         }
       }
   }
-
-  // New light jobs feed available only in dev env
-  def lightJobs = Action {
-    implicit request =>
-      LightJobsAgent.matchingAds(segment) match {
-        case Nil => JsonNotFound.apply()
-        case jobs => {
-          val shuffled = Random.shuffle(jobs)
-          Cached(60)(JsonComponent(views.html.lightjobs(shuffled take 5)))
-        }
-      }
-  }
-
 }
