+++ commercial/app/controllers/commercial/MasterClasses.scala
@@ -3,7 +3,6 @@ package controllers.commercial
 import play.api.mvc._
 import common.{JsonNotFound, JsonComponent}
 import model.commercial.masterclasses.MasterClassAgent
-import scala.util.Random
 import model.Cached
 
 object MasterClasses extends Controller {
@@ -13,8 +12,7 @@ object MasterClasses extends Controller {
       MasterClassAgent.getUpcoming match {
         case Nil => JsonNotFound.apply()
         case upcoming => {
-          val shuffled = Random.shuffle(upcoming)
-          Cached(60)(JsonComponent(views.html.masterclasses(shuffled take 3)))
+          Cached(60)(JsonComponent(views.html.masterclasses(upcoming take 3)))
         }
       }
   }
