+++ commercial/app/model/commercial/AdAgent.scala
@@ -1,6 +1,7 @@
 package model.commercial
 
 import common.AkkaAgent
+import scala.util.Random
 
 trait AdAgent[T <: Ad] {
 
@@ -10,10 +11,10 @@ trait AdAgent[T <: Ad] {
 
   protected def updateCurrentAds(ads: Seq[T]) = agent send ads
 
-  def matchingAds(segment: Segment, adsToChooseFrom: Seq[T] = currentAds): Seq[T] = {
+  def adsTargetedAt(segment: Segment, adsToChooseFrom: Seq[T] = currentAds): Seq[T] = {
     // TODO: reinstate repeatVisitor condition when the time is right
     //adsToChooseFrom filter (segment.isRepeatVisitor && _.matches(segment))
-    adsToChooseFrom filter (_.isTargetedAt(segment))
+    Random.shuffle(adsToChooseFrom filter (_.isTargetedAt(segment)))
   }
 
   def stop() {
