+++ commercial/app/model/commercial/jobs/JobsApi.scala
@@ -1,57 +1,37 @@
 package model.commercial.jobs
 
 import scala.concurrent.Future
-import org.joda.time.format.DateTimeFormat
 import scala.xml.Elem
 import conf.CommercialConfiguration
 import model.commercial.Utils.OptString
 import model.commercial.XmlAdsApi
+import org.joda.time.format.DateTimeFormat
 
 object JobsApi extends XmlAdsApi[Job] {
 
   protected val adTypeName = "Jobs"
 
-  private lazy val optUrl = CommercialConfiguration.getProperty("jobs.api.url")
-  private lazy val optKey = CommercialConfiguration.getProperty("jobs.api.key")
+  private val feedDate = DateTimeFormat.forPattern("yyyy-MM-dd").print(System.currentTimeMillis)
+  private lazy val urlTemplate = CommercialConfiguration.getProperty("jobs.api.url.template")
+  private lazy val url = urlTemplate map (_ replace("yyyy-MM-dd", feedDate))
 
-  override protected val loadTimeout = 60000
+  override protected val loadTimeout = 20000
 
-  private val dateFormat = DateTimeFormat.forPattern("dd/MM/yyyy HH:mm:ss")
-
-  override def cleanResponseBody(body: String) = body.replace(0x001b.toChar, ' ')
-
-  private def loadJobs: Future[Seq[Job]] = loadAds {
-    for {
-      url <- optUrl
-      key <- optKey
-    } yield s"$url?login=$key"
-  }
+  override def cleanResponseBody(body: String) = body.dropWhile(_ != '<')
 
   def parse(xml: Elem): Seq[Job] = {
     (xml \ "Job") map {
       job =>
         Job(
           (job \ "JobID").text.toInt,
-          (job \ "AdType").text,
-          dateFormat.parseDateTime((job \ "StartDateTime").text),
-          dateFormat.parseDateTime((job \ "EndDateTime").text),
-          (job \ "IsPremium").text.toBoolean,
-          (job \ "PositionType").text,
           (job \ "JobTitle").text,
           (job \ "ShortJobDescription").text,
-          (job \ "SalaryDescription").text,
-          OptString((job \ "LocationDescription").text),
+          (job \ "RecruiterName").text,
           OptString((job \ "RecruiterLogoURL").text),
-          OptString((job \ "EmployerLogoURL").text),
-          (job \ "JobListingURL").text,
-          (job \ "ApplyURL").text,
-          ((job \ "Sector" \ "Description") map (_.text)).distinct,
-          (job \ "Location" \ "Description") map (_.text)
+          ((job \ "Sectors" \ "Sector") map (_.text.toInt)).toSet
         )
     }
   }
 
-  def getCurrentJobs(allJobs: Future[Seq[Job]] = loadJobs): Future[Seq[Job]] = {
-    allJobs map (_ filter (_.isCurrent))
-  }
+  def getJobs: Future[Seq[Job]] = loadAds(url)
 }
