--- commercial/app/views/lightjobs.scala.html
@@ -1,20 +1,0 @@
-@(jobs: Seq[model.commercial.jobs.LightJob])
-
-<div class="commercial commercial--masterclasses">
-    <div class="commercial__head">
-        <h3 class="commercial__title">guardian<span>jobs</span></h3>
-    </div>
-    <div class="commercial__body">
-        <table>
-            @for(job <- jobs) {
-                <tr>
-                    <td>@job.recruiterLogoUrl.map{ url => <img src="@url" />}</td>
-                    <td>@job.id</td>
-                    <td><a href="@job.listingUrl">@job.title</a></td>
-                    <td>@(job.sectorIds.mkString("; "))</td>
-                </tr>
-            }
-        </table>
-    </div>
-    <div class="commercial__foot"></div>
-</div>
