+++ commercial/test/model/commercial/jobs/Fixtures.scala
@@ -1,165 +1,100 @@
 package model.commercial.jobs
 
-import org.joda.time.DateTime
 
 object Fixtures {
 
   val xml =
     """
       |<Jobs>
-      |<Job><ClientName><![CDATA[MACMILLAN CANCER SUPPORT]]></ClientName><JobID><![CDATA[4721342]]></JobID><JobTitle><![CDATA[Macmillan Cancer Support Social Care Programme Research & Communications Intern]]></JobTitle><JobDescription><![CDATA[<p><b>Why do you need me? </b></p>
-      |<p>You will be working as part of the busy Social Care Programme within the dynamic Health and Social Care Team at Macmillan. The team is part of the Cancer Services Innovation Directorate. The Social Care Programme is working on number of exciting new initiatives which aims to improve Social Care support for people with cancer.</p>
-      |<p>Our team needs an enthusiastic volunteer intern to provide project support to the Social Care Programme. We need someone to utilise the skills they already have to aid the progress of the ongoing projects within the programme.</p>
-      |<p>You will also be given the opportunity to develop your own project planning and communication skills. This role is also a fantastic opportunity to develop your team work and event management skills and experience working on innovative projects.</p>
-      |<p><b>What activities will I be involved in? </b></p>
-      |<ul>
-      |    <li>Completing a research project of Macmillan Grants to gather evidence of what PABC are asking for and therefore what their social care needs are- this will be qualitative and quantitative research.</li>
-      |    <li>Creating databases and spreadsheets</li>
-      |    <li>Completing scoping work within the field</li>
-      |    <li>Gathering case studies</li>
-      |    <li>Interviewing volunteers about their experience and gathering data</li>
-      |    <li>Work with internal teams in communications, marketing and media</li>
-      |    <li>Helping to develop resources</li>
-      |    <li>Producing power point presentations, briefing documents, learning resources e.t.c</li>
-      |</ul>
-      |<p><b>What skills and abilities will I need to have? </b></p>
-      |<ul>
-      |    <li>Strong research and analysis skills</li>
-      |    <li>Experience with working with large amounts of data</li>
-      |    <li>Strong organisational skills with a proven ability to meet deadlines and attention to detail</li>
-      |    <li>Excellent written communication skills and a confident user of PowerPoint and Word and Excel</li>
-      |    <li>Ability to use own initiative within specified guidelines to aid the progress of projects</li>
-      |    <li>A demonstrable ability to work as part of a team</li>
-      |    <li>Motivation and enthusiasm to see the project through from start to finish</li>
-      |    <li>Enthusiasm for the cause and ability to represent Macmillan Cancer Support</li>
-      |</ul>
-      |<p><b>What are the goals? </b></p>
-      |<ul>
-      |    <li>Supporting the Social Care Programme
-      |    <ul>
-      |        <li>Providing project support to the Programme lead and Project Officer</li>
-      |        <li>Supporting and developing the case studies and the Macmillan grant research projects within the Programme.</li>
-      |    </ul>
-      |    </li>
-      |    <li>Supporting Macmillan by being an ambassador and representative of Macmillan&rsquo;s internship scheme
-      |    <ul>
-      |        <li>inspire and engage supporters to give and do more</li>
-      |    </ul>
-      |    </li>
-      |</ul>
-      |<p><b>What are the benefits to me? </b></p>
-      |<p>This is a unique opportunity for those wanting to develop and improve project planning skills and for those wanting to move into charity sector. You will:</p>
-      |<ul>
-      |    <li>Gain hands-on experience of working in the UK&rsquo;s largest cancer support charity</li>
-      |    <li>Develop essentials skills such as: communication, organisation, project planning and team working</li>
-      |    <li>experiencing working with media, communications and marketing</li>
-      |    <li>In-depth knowledge of Health and Social Care across the UK and what the Social Care Programme are doing to improve the lives of those affected by cancer in this field</li>
-      |    <li>Be part of a supportive, lively and fun team who will facilitate your development and offer support</li>
-      |    <li>Enhance your CV and gain written references at the end of your internship</li>
-      |</ul>
-      |<p><b>How much time should I offer? </b></p>
-      |<ul>
-      |    <li>This internship is available for 3-5 days a week from for a three month period starting as soon as possible. Office hours are 9am - 5pm Monday to Thursday and 9am - 4.30 pm on Friday.</li>
-      |    <li>We can be flexible with days and times to suit your requirements.</li>
-      |</ul>
-      |<p><b>Where will I be based? Will I need my own transport? </b></p>
-      |<ul>
-      |    <li>UKO Macmillan Cancer Support, 89 Albert Embankment, SE1 7UQ.</li>
-      |    <li>No specific transport will be required.</li>
-      |</ul>
-      |<p><b>Is there an induction and training? </b></p>
-      |<p>We will ensure you have the training and information you need through:</p>
-      |<ul>
-      |    <li>Our comprehensive induction pack</li>
-      |    <li>Any further on-the-job coaching of organisational process needed for the role</li>
-      |</ul>
-      |<p><b>What ongoing support/guidance will there be? </b></p>
-      |<p>You will be fully supported by Fiona Smith the Social Care Project Officer for the Social Care Programme and the Health and Social Care team throughout your time with us. Including:</p>
-      |<ul>
-      |    <li>Regular meetings with Fiona and opportunities for feedback and coaching</li>
-      |    <li>Mentoring and CV advice should you want it</li>
-      |    <li>Opportunities to network with staff, volunteers and other Interns during the placement</li>
-      |</ul>
-      |<p><b>Will my expenses be paid? </b></p>
-      |<p>The internship is unpaid but we pay volunteers&rsquo; &ldquo;out-of-pocket&rdquo; expenses within agreed guidelines which should cover your lunch and travel.</p>
-      |<p><b>What&rsquo;s the next step? </b></p>
-      |<p>You need to fill out an application form and send it to us. We will acknowledge your application and let you know whether you have been shortlisted for interview.</p>]]></JobDescription><ShortJobDescription><![CDATA[We need an enthusiastic volunteer intern to provide project support to the busy Social Care Programme (SCP) within the dynamic Health and Social Care Team at Macmillan.]]></ShortJobDescription><JobRef><![CDATA[]]></JobRef><JobPostCode><![CDATA[SE1 7UQ]]></JobPostCode><LocationDescription><![CDATA[Vauxhall, London]]></LocationDescription><SalaryDescription><![CDATA[Unpaid Voluntary Work]]></SalaryDescription><Sector><Description><![CDATA[Charities]]></Description><Description><![CDATA[Health]]></Description><Description><![CDATA[Health]]></Description><Description><![CDATA[Social care]]></Description><Description><![CDATA[Policy & research]]></Description><Description><![CDATA[Social welfare]]></Description><Description><![CDATA[Academic & research]]></Description><Description><![CDATA[Support services]]></Description><Description><![CDATA[Adult Care]]></Description><Description><![CDATA[Family Support]]></Description><Description><![CDATA[Health]]></Description></Sector><Location><Description><![CDATA[UK]]></Description><Description><![CDATA[England]]></Description><Description><![CDATA[Greater London]]></Description><Description><![CDATA[South London]]></Description></Location><SalaryRange /><PositionType><![CDATA[Temp]]></PositionType><Hours><![CDATA[Part Time]]></Hours><StartDateTime><![CDATA[10/10/2013 08:00:00]]></StartDateTime><EndDateTime><![CDATA[10/11/2113 23:59:00]]></EndDateTime><ContactTitle><![CDATA[]]></ContactTitle><ContactFirstName><![CDATA[]]></ContactFirstName><ContactLastName><![CDATA[]]></ContactLastName><Address1><![CDATA[]]></Address1><Address2><![CDATA[]]></Address2><Address3><![CDATA[]]></Address3><CityTown><![CDATA[]]></CityTown><County><![CDATA[]]></County><PostCode><![CDATA[]]></PostCode><Country><![CDATA[]]></Country><Email><![CDATA[]]></Email><Telephone><![CDATA[]]></Telephone><Fax><![CDATA[]]></Fax><ApplicationMethod><![CDATA[ExternalRedirect]]></ApplicationMethod><PreferredApplicationMethod><![CDATA[ExternalRedirect]]></PreferredApplicationMethod><ExternalApplicationURL><![CDATA[http://volunteering.macmillan.org.uk/Opportunity/Details/1753]]></ExternalApplicationURL><AdType><![CDATA[Volunteering]]></AdType><JobListingURL><![CDATA[http://jobs.theguardian.com/job/4721342/macmillan-cancer-support-social-care-programme-research-and-communications-intern/?TrackID=554176]]></JobListingURL><ApplyURL><![CDATA[http://jobs.theguardian.com/apply/4721342/macmillan-cancer-support-social-care-programme-research-and-communications-intern/?TrackID=554176]]></ApplyURL><IsPremium><![CDATA[false]]></IsPremium><RecruiterLogoURL><![CDATA[http://jobs.theguardian.com/getasset/?uiAssetID=1f8c821a-6de6-4f5d-ac32-fee9b8fb0b42]]></RecruiterLogoURL><EmployerLogoURL><![CDATA[]]></EmployerLogoURL><SecondLogoURL><![CDATA[]]></SecondLogoURL></Job>
-      |<Job><ClientName><![CDATA[HERTFORDSHIRE COUNTY COUNCIL]]></ClientName><JobID><![CDATA[4726895]]></JobID><JobTitle><![CDATA[ARC Worker]]></JobTitle><JobDescription><![CDATA[<p>We appreciate that not every teenager wants our help. But when they need a helping hand in life we hope we can offer them the support they need. The Adolescent Resource Centre is has been running since September 2009, offering a countywide service to children, young people and their families. We provide a 24 hour, seven days a week, flexible service which supports placements, and our own specialist foster carers. Experience in social care goes without saying but whatever the level you apply for the one thing you must have is an absolute commitment to improving the lives of children.</p>
-      |
-      |<p>The ARC Workers are part of a specialist multi disciplinary team, putting in place intensive care packages to provide holistic support to young people who have experienced trauma in their life. Our staff have been trained to work to the Attachment, Self-Regulation and Competency model, which equips them with the skills needed to work with young people who have complex needs, helping them to build secure attachments and become self-reliant. It is essential all our staff are able to work shifts, some weekends and bank holidays and experience of working with children, young people and families is a must, as well as the commitment to make a difference. So, join our team and receive a full induction and specialist training and help make a difference.</p>
-      |
-      |<p>For an informal discussion please contact Steve Gentry on 01707 343630.</p>
-      |
-      |<p>To apply, please register and submit your application via the Council&rsquo;s Talentlink recruitment system http://www.hertsdirect.org/actweb/talentlink/. The job reference number is <b>HCC01236</b>. The system will allow you to apply via on-line application where you should complete a supporting statement demonstrating how you meet the requirements for the role.</p>
-      |
-      |<p><b>Closing Date: 31st October 2013</b></p>]]></JobDescription><ShortJobDescription><![CDATA[We appreciate that not every teenager wants our help. But when they need a helping hand in life we hope we can offer them the support they need. The Adolescent Resource Centre is has been running since September 2009.]]></ShortJobDescription><JobRef><![CDATA[HCC01236]]></JobRef><JobPostCode><![CDATA[]]></JobPostCode><LocationDescription><![CDATA[Countywide]]></LocationDescription><SalaryDescription><![CDATA[£18,453 - £24,646 PA Inclusive]]></SalaryDescription><Sector><Description><![CDATA[Social care]]></Description><Description><![CDATA[Children]]></Description><Description><![CDATA[Youth]]></Description><Description><![CDATA[Family Support]]></Description></Sector><Location><Description><![CDATA[UK]]></Description><Description><![CDATA[England]]></Description><Description><![CDATA[South East]]></Description><Description><![CDATA[Hertfordshire]]></Description></Location><SalaryRange><Description><![CDATA[£15,000 - £20,000]]></Description><Description><![CDATA[£20,000 - £25,000]]></Description></SalaryRange><PositionType><![CDATA[Permanent]]></PositionType><Hours><![CDATA[Full Time]]></Hours><StartDateTime><![CDATA[18/10/2013 08:00:00]]></StartDateTime><EndDateTime><![CDATA[31/10/2113 23:59:00]]></EndDateTime><ContactTitle><![CDATA[]]></ContactTitle><ContactFirstName><![CDATA[]]></ContactFirstName><ContactLastName><![CDATA[]]></ContactLastName><Address1><![CDATA[]]></Address1><Address2><![CDATA[]]></Address2><Address3><![CDATA[]]></Address3><CityTown><![CDATA[]]></CityTown><County><![CDATA[]]></County><PostCode><![CDATA[]]></PostCode><Country><![CDATA[]]></Country><Email><![CDATA[]]></Email><Telephone><![CDATA[]]></Telephone><Fax><![CDATA[]]></Fax><ApplicationMethod><![CDATA[ExternalRedirect]]></ApplicationMethod><PreferredApplicationMethod><![CDATA[ExternalRedirect]]></PreferredApplicationMethod><ExternalApplicationURL><![CDATA[http://www.hertsdirect.org/actweb/talentlink/details.cfm?nPostingID=1170&nPostingTargetID=1908&option=52&sort=DESC&respnr=1&ID=QY6FK026203F3VBQBV7V4F6GJ&JobNum=HCC01236&Resultsperpage=10&lg=UK&mask=external]]></ExternalApplicationURL><AdType><![CDATA[Job vacancy]]></AdType><JobListingURL><![CDATA[http://jobs.theguardian.com/job/4726895/arc-worker/?TrackID=554176]]></JobListingURL><ApplyURL><![CDATA[http://jobs.theguardian.com/apply/4726895/arc-worker/?TrackID=554176]]></ApplyURL><IsPremium><![CDATA[false]]></IsPremium><RecruiterLogoURL><![CDATA[http://jobs.theguardian.com/getasset/?uiAssetID=bfe4737e-06a3-4688-aa9c-aa4e952a4795]]></RecruiterLogoURL><EmployerLogoURL><![CDATA[]]></EmployerLogoURL><SecondLogoURL><![CDATA[]]></SecondLogoURL></Job>
-      |<Job><ClientName><![CDATA[TARGETJOBS.CO.UK]]></ClientName><JobID><![CDATA[4729001]]></JobID><JobTitle><![CDATA[Graduate PR, Communications & Public Affairs]]></JobTitle><JobDescription><![CDATA[<p>An ever-evolving heritage. A truly exciting future. Home to two of the world&rsquo;s most iconic brands. At Jaguar Land Rover we are continually redefining the benchmark for excellence and setting the standards that others want to follow. And with increasing demand and even greater ambitions, finding the future leaders of our business has never been more important than now. That&rsquo;s where you join the journey.</p>
-      |<p>PR, Communications &amp; Public Affairs is the voice of our rapidly expanding global business. Join us on our two-year Graduate programme, and you&rsquo;ll play a key role in helping to deliver Jaguar Land Rover&rsquo;s global PR strategy in over 170 markets worldwide.</p>
-      |<p>Media-savvy and one step ahead, you&rsquo;ll help drive conversation and ensure the exposure of our business across multiple channels in the UK and abroad. You&rsquo;ll protect and promote our iconic brands, generating excitement for our premium products in all sectors. And you&rsquo;ll make your impact on a diverse range of projects, from writing global press releases and managing relationships with the media, to working with creative agencies and planning world-wide product launches.</p>
-      |<p>Your journey will begin with a thorough introduction to both brands, followed by carefully planned development throughout the programme. You&rsquo;ll work with industry-leading experts to develop specialist skills and you&rsquo;ll be supported to gain professional qualifications and accreditation. At every stage, we will give you the encouragement and freedom to achieve your full potential, drive your own success and enhance Jaguar Land Rover&rsquo;s outstanding reputation across the globe.</p>
-      |<p>As you&rsquo;d expect from one of the world's most revered organisations, an outstanding range of rewards and benefits await those who have the vision and drive to continue our global success &ndash; including a competitive starting salary, joining bonus, pension scheme and discounted car purchase scheme. All this and more makes Jaguar Land Rover an excellent place to start your journey and build a world-class career.</p>]]></JobDescription><ShortJobDescription><![CDATA[An ever-evolving heritage. A truly exciting future. Home to two of the world&rsquo;s most iconic brands. At Jaguar Land Rover we are continually redefining the benchmark for excellence and setting the standards that others want to follow. And with increas]]></ShortJobDescription><JobRef><![CDATA[329049 2013-10-03T17:15:53.000+01:00]]></JobRef><JobPostCode><![CDATA[]]></JobPostCode><LocationDescription><![CDATA[]]></LocationDescription><SalaryDescription><![CDATA[Competitive]]></SalaryDescription><Sector><Description><![CDATA[Marketing & PR]]></Description><Description><![CDATA[Digital Marketing]]></Description></Sector><Location><Description><![CDATA[UK]]></Description></Location><SalaryRange /><PositionType><![CDATA[Permanent]]></PositionType><Hours><![CDATA[Full Time]]></Hours><StartDateTime><![CDATA[24/09/2013 00:03:00]]></StartDateTime><EndDateTime><![CDATA[24/10/2113 23:59:59]]></EndDateTime><ContactTitle><![CDATA[]]></ContactTitle><ContactFirstName><![CDATA[]]></ContactFirstName><ContactLastName><![CDATA[]]></ContactLastName><Address1><![CDATA[]]></Address1><Address2><![CDATA[]]></Address2><Address3><![CDATA[]]></Address3><CityTown><![CDATA[]]></CityTown><County><![CDATA[]]></County><PostCode><![CDATA[]]></PostCode><Country><![CDATA[]]></Country><Email><![CDATA[]]></Email><Telephone><![CDATA[]]></Telephone><Fax><![CDATA[]]></Fax><ApplicationMethod><![CDATA[ExternalRedirect]]></ApplicationMethod><PreferredApplicationMethod><![CDATA[ExternalRedirect]]></PreferredApplicationMethod><ExternalApplicationURL><![CDATA[http://targetjobs.co.uk/employer-hubs/jaguar-land-rover/329049-graduate-pr-communications-public-affairs]]></ExternalApplicationURL><AdType><![CDATA[Graduate scheme]]></AdType><JobListingURL><![CDATA[http://jobs.theguardian.com/job/4729001/graduate-pr-communications-and-public-affairs/?TrackID=554176]]></JobListingURL><ApplyURL><![CDATA[http://jobs.theguardian.com/apply/4729001/graduate-pr-communications-and-public-affairs/?TrackID=554176]]></ApplyURL><IsPremium><![CDATA[false]]></IsPremium><RecruiterLogoURL><![CDATA[]]></RecruiterLogoURL><EmployerLogoURL><![CDATA[]]></EmployerLogoURL><SecondLogoURL><![CDATA[]]></SecondLogoURL></Job>
-      |<Job><ClientName><![CDATA[TARGETJOBS.CO.UK]]></ClientName><JobID><![CDATA[4729001]]></JobID><JobTitle><![CDATA[Graduate PR, Communications & Public Affairs]]></JobTitle><JobDescription><![CDATA[<p>An ever-evolving heritage. A truly exciting future. Home to two of the world&rsquo;s most iconic brands. At Jaguar Land Rover we are continually redefining the benchmark for excellence and setting the standards that others want to follow. And with increasing demand and even greater ambitions, finding the future leaders of our business has never been more important than now. That&rsquo;s where you join the journey.</p>
-      |<p>PR, Communications &amp; Public Affairs is the voice of our rapidly expanding global business. Join us on our two-year Graduate programme, and you&rsquo;ll play a key role in helping to deliver Jaguar Land Rover&rsquo;s global PR strategy in over 170 markets worldwide.</p>
-      |<p>Media-savvy and one step ahead, you&rsquo;ll help drive conversation and ensure the exposure of our business across multiple channels in the UK and abroad. You&rsquo;ll protect and promote our iconic brands, generating excitement for our premium products in all sectors. And you&rsquo;ll make your impact on a diverse range of projects, from writing global press releases and managing relationships with the media, to working with creative agencies and planning world-wide product launches.</p>
-      |<p>Your journey will begin with a thorough introduction to both brands, followed by carefully planned development throughout the programme. You&rsquo;ll work with industry-leading experts to develop specialist skills and you&rsquo;ll be supported to gain professional qualifications and accreditation. At every stage, we will give you the encouragement and freedom to achieve your full potential, drive your own success and enhance Jaguar Land Rover&rsquo;s outstanding reputation across the globe.</p>
-      |<p>As you&rsquo;d expect from one of the world's most revered organisations, an outstanding range of rewards and benefits await those who have the vision and drive to continue our global success &ndash; including a competitive starting salary, joining bonus, pension scheme and discounted car purchase scheme. All this and more makes Jaguar Land Rover an excellent place to start your journey and build a world-class career.</p>]]></JobDescription><ShortJobDescription><![CDATA[An ever-evolving heritage. A truly exciting future. Home to two of the world&rsquo;s most iconic brands. At Jaguar Land Rover we are continually redefining the benchmark for excellence and setting the standards that others want to follow. And with increas]]></ShortJobDescription><JobRef><![CDATA[329049 2013-10-03T17:15:53.000+01:00]]></JobRef><JobPostCode><![CDATA[]]></JobPostCode><LocationDescription><![CDATA[]]></LocationDescription><SalaryDescription><![CDATA[Competitive]]></SalaryDescription><Sector><Description><![CDATA[Marketing & PR]]></Description><Description><![CDATA[Digital Marketing]]></Description></Sector><Location><Description><![CDATA[UK]]></Description></Location><SalaryRange /><PositionType><![CDATA[Permanent]]></PositionType><Hours><![CDATA[Full Time]]></Hours><StartDateTime><![CDATA[24/09/2013 00:03:00]]></StartDateTime><EndDateTime><![CDATA[24/10/2013 23:59:59]]></EndDateTime><ContactTitle><![CDATA[]]></ContactTitle><ContactFirstName><![CDATA[]]></ContactFirstName><ContactLastName><![CDATA[]]></ContactLastName><Address1><![CDATA[]]></Address1><Address2><![CDATA[]]></Address2><Address3><![CDATA[]]></Address3><CityTown><![CDATA[]]></CityTown><County><![CDATA[]]></County><PostCode><![CDATA[]]></PostCode><Country><![CDATA[]]></Country><Email><![CDATA[]]></Email><Telephone><![CDATA[]]></Telephone><Fax><![CDATA[]]></Fax><ApplicationMethod><![CDATA[ExternalRedirect]]></ApplicationMethod><PreferredApplicationMethod><![CDATA[ExternalRedirect]]></PreferredApplicationMethod><ExternalApplicationURL><![CDATA[http://targetjobs.co.uk/employer-hubs/jaguar-land-rover/329049-graduate-pr-communications-public-affairs]]></ExternalApplicationURL><AdType><![CDATA[Graduate scheme]]></AdType><JobListingURL><![CDATA[http://jobs.theguardian.com/job/4729001/graduate-pr-communications-and-public-affairs/?TrackID=554176]]></JobListingURL><ApplyURL><![CDATA[http://jobs.theguardian.com/apply/4729001/graduate-pr-communications-and-public-affairs/?TrackID=554176]]></ApplyURL><IsPremium><![CDATA[false]]></IsPremium><RecruiterLogoURL><![CDATA[]]></RecruiterLogoURL><EmployerLogoURL><![CDATA[]]></EmployerLogoURL><SecondLogoURL><![CDATA[]]></SecondLogoURL></Job>
-      |      |</Jobs>
+      |  <Job>
+      |    <JobID>1058606</JobID>
+      |    <JobTitle>Area Management Training Programme (Graduate Area Manager)</JobTitle>
+      |    <RecruiterName>ALDI</RecruiterName>
+      |    <ShortJobDescription>We're only looking for outstanding individuals for the Aldi Management Programme.</ShortJobDescription>
+      |    <Sectors>
+      |      <Sector>149</Sector>
+      |      <Sector>158</Sector>
+      |      <Sector>245</Sector>
+      |    </Sectors>
+      |    <RecruiterLogoURL>http://jobs.theguardian.com/getasset/?uiAssetID=833D7672-6B21-4D81-BCE1-9CEF11CCEA21</RecruiterLogoURL>
+      |  </Job>
+      |  <Job>
+      |    <JobID>4302057</JobID>
+      |    <JobTitle>Listen Laugh and Learn in Worcestershire</JobTitle>
+      |    <RecruiterName>MOTOR NEURONE DISEASE ASSOCIATION</RecruiterName>
+      |    <ShortJobDescription>The role of Association Visitor is a highly rewarding one with many opportunities for developing your own skills and knowledge, whilst being part of a team. Professional qualifications or personal experience of MND are not necessary...</ShortJobDescription>
+      |    <Sectors>
+      |      <Sector>111</Sector>
+      |      <Sector>112</Sector>
+      |      <Sector>121</Sector>
+      |      <Sector>286</Sector>
+      |      <Sector>600117</Sector>
+      |      <Sector>600190</Sector>
+      |    </Sectors>
+      |    <RecruiterLogoURL>http://jobs.theguardian.com/getasset/?uiAssetID=70666A14-BD49-4BB8-AC6A-FCF8B716E131</RecruiterLogoURL>
+      |  </Job>
+      |  <Job>
+      |    <JobID>4365671</JobID>
+      |    <JobTitle>Female Youth Work Volunteer in sports</JobTitle>
+      |    <RecruiterName>CATCH 22</RecruiterName>
+      |    <ShortJobDescription>The Active Women's project helps young women access sports activities in their local area in Southampton. Volunteers will support the running of these sessions, as well as supporting the young women in accessing the sessions.</ShortJobDescription>
+      |    <Sectors>
+      |      <Sector>111</Sector>
+      |      <Sector>112</Sector>
+      |      <Sector>115</Sector>
+      |      <Sector>219</Sector>
+      |      <Sector>222</Sector>
+      |      <Sector>600118</Sector>
+      |      <Sector>600119</Sector>
+      |    </Sectors>
+      |    <RecruiterLogoURL>http://jobs.theguardian.com/getasset/?uiAssetID=1D428A1F-65B4-4ADE-8CA3-91F2AE7423E3</RecruiterLogoURL>
+      |  </Job>
+      |  <Job>
+      |    <JobID>4411510</JobID>
+      |    <JobTitle>Air Traffic Controller</JobTitle>
+      |    <RecruiterName>RAF CAREERS</RecruiterName>
+      |    <ShortJobDescription>Control some of the world’s most modern aircraft using radar and communications equipment</ShortJobDescription>
+      |    <Sectors>
+      |      <Sector>166</Sector>
+      |      <Sector>308</Sector>
+      |    </Sectors>
+      |    <RecruiterLogoURL>http://jobs.theguardian.com/getasset/?uiAssetID=14357B35-BBA3-4F96-ADF3-F9ABC82C894C</RecruiterLogoURL>
+      |  </Job>
+      |</Jobs>
       | """.stripMargin
 
   val jobs = List(
 
-    Job(4721342,
-      "Volunteering",
-      new DateTime("2013-10-10T08:00:00.000+01:00"),
-      new DateTime("2113-11-10T23:59:00.000Z"),
-      isPremium = false,
-      "Temp",
-      "Macmillan Cancer Support Social Care Programme Research & Communications Intern",
-      "We need an enthusiastic volunteer intern to provide project support to the busy Social Care Programme (SCP) within the dynamic Health and Social Care Team at Macmillan.",
-      "Unpaid Voluntary Work",
-      Some("Vauxhall, London"),
-      Some("http://jobs.theguardian.com/getasset/?uiAssetID=1f8c821a-6de6-4f5d-ac32-fee9b8fb0b42"),
-      None,
-      "http://jobs.theguardian.com/job/4721342/macmillan-cancer-support-social-care-programme-research-and-communications-intern/?TrackID=554176",
-      "http://jobs.theguardian.com/apply/4721342/macmillan-cancer-support-social-care-programme-research-and-communications-intern/?TrackID=554176",
-      List("Charities", "Health", "Social care", "Policy & research", "Social welfare", "Academic & research", "Support services", "Adult Care", "Family Support"),
-      List("UK", "England", "Greater London", "South London")
+    Job(1058606,
+      "Area Management Training Programme (Graduate Area Manager)",
+      "We're only looking for outstanding individuals for the Aldi Management Programme.",
+      "ALDI",
+      Some("http://jobs.theguardian.com/getasset/?uiAssetID=833D7672-6B21-4D81-BCE1-9CEF11CCEA21"),
+      Set(149, 158, 245)
     ),
 
-    Job(4726895,
-      "Job vacancy",
-      new DateTime("2013-10-18T08:00:00.000+01:00"),
-      new DateTime("2113-10-31T23:59:00.000Z"),
-      isPremium = false,
-      "Permanent",
-      "ARC Worker",
-      "We appreciate that not every teenager wants our help. But when they need a helping hand in life we hope we can offer them the support they need. The Adolescent Resource Centre is has been running since September 2009.",
-      "£18,453 - £24,646 PA Inclusive",
-      Some("Countywide"),
-      Some("http://jobs.theguardian.com/getasset/?uiAssetID=bfe4737e-06a3-4688-aa9c-aa4e952a4795"),
-      None,
-      "http://jobs.theguardian.com/job/4726895/arc-worker/?TrackID=554176",
-      "http://jobs.theguardian.com/apply/4726895/arc-worker/?TrackID=554176",
-      List("Social care", "Children", "Youth", "Family Support"),
-      List("UK", "England", "South East", "Hertfordshire")
+    Job(4302057,
+      "Listen Laugh and Learn in Worcestershire",
+      "The role of Association Visitor is a highly rewarding one with many opportunities for developing your own skills and knowledge, whilst being part of a team. Professional qualifications or personal experience of MND are not necessary...",
+      "MOTOR NEURONE DISEASE ASSOCIATION",
+      Some("http://jobs.theguardian.com/getasset/?uiAssetID=70666A14-BD49-4BB8-AC6A-FCF8B716E131"),
+      Set(111, 112, 121, 286, 600117, 600190)
     ),
 
-    Job(4729001,
-      "Graduate scheme",
-      new DateTime("2013-09-24T00:03:00.000+01:00"),
-      new DateTime("2113-10-24T23:59:59.000+01:00"),
-      isPremium = false,
-      "Permanent",
-      "Graduate PR, Communications & Public Affairs",
-      "An ever-evolving heritage. A truly exciting future. Home to two of the world&rsquo;s most iconic brands. At Jaguar Land Rover we are continually redefining the benchmark for excellence and setting the standards that others want to follow. And with increas",
-      "Competitive",
-      None,
-      None,
-      None,
-      "http://jobs.theguardian.com/job/4729001/graduate-pr-communications-and-public-affairs/?TrackID=554176",
-      "http://jobs.theguardian.com/apply/4729001/graduate-pr-communications-and-public-affairs/?TrackID=554176",
-      List("Marketing & PR", "Digital Marketing"),
-      List("UK")
+    Job(4365671,
+      "Female Youth Work Volunteer in sports",
+      "The Active Women's project helps young women access sports activities in their local area in Southampton. Volunteers will support the running of these sessions, as well as supporting the young women in accessing the sessions.",
+      "CATCH 22",
+      Some("http://jobs.theguardian.com/getasset/?uiAssetID=1D428A1F-65B4-4ADE-8CA3-91F2AE7423E3"),
+      Set(111, 112, 115, 219, 222, 600118, 600119)
+    ),
+
+    Job(4411510,
+      "Air Traffic Controller",
+      "Control some of the world’s most modern aircraft using radar and communications equipment",
+      "RAF CAREERS",
+      Some("http://jobs.theguardian.com/getasset/?uiAssetID=14357B35-BBA3-4F96-ADF3-F9ABC82C894C"),
+      Set(166, 308)
     )
 
   )
